Add "Vintage Stock Photos" to free images
We've all been there. An old photo - family snaphot or thrift store discard - catches your eye and you can't look away. The image is technically a hot mess but there's an essential, visceral appeal about the subject - a faded person or place that unexpectedly and elegantly captures a lost moment. This happens a lot around here and we fall in love with the blurry, scratched, grainy, misfit images depicting life in the middle of the 20th century.

We started Vintage Stock Photos to sell some of these great images. People loved the images, but sales were slow and it's hard to be a niche stock site, so we decided to give away the image collection for free.

About image production

Our photos come from collections we've acquired, and we own the original transparencies for all images. Transparencies are scanned, then post processed and keyworded to the best of our abilities. We archive the original scan and output a resized jpg file for distribution. If you feel the version we've posted it too small or too processed, let us know and we may be able to rescan, resize or reprocess the image.

If you have any other questions, please use the form in the Support section and we'll get back to you ASAP.

Thanks,
Vintage Stock Photos