Add Open Device Labs
The fragmentation of operating systems and browsers as well as the diversity of internet-enabled devices makes it impossible for the vast majority of developers to personally own a representative pool of test devices. Nevertheless, quality assurance across real devices is a must to ensure a pleasant user experience, sufficient stability and security.

Open Device Labs (ODLs, #ODL) are a grass roots community movement. They establish shared community pools of internet connected devices for testing purposes of web and app developers. In result, ODLs lead to an ultimate improvement of the web & app experience both for developers and for consumers.

OpenDeviceLab.com is here to serve three major goals:

* help people to locate the right Open Device Lab for the job
* explain and promote the Open Device Lab movement - and ODLs to become more visible
* attract Contributors and Sponsors to help and donate to ODLs.